extract person class
